FE_221029_board : 전체화면 api 연동 완료

+++ android/app/src/main/java/com/ssafy/zip/android/data/User.kt
@@ -1,4 +1,5 @@
 package com.ssafy.zip.android.data
+
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
 import kotlinx.parcelize.Parcelize
@@ -8,7 +9,7 @@ data class User(
     @SerializedName("familyResponseDTO")
     var family : Family,
     var hasFamily : Boolean,
-    var id : Int,
+    var id : Long,
     var name : String,
     var nickname: String,
     var profileImg : String? // 추후 수정?
